+++ include/functions.h
@@ -182,7 +182,7 @@ void particleEmitter_setSprite(ParticleEmitter *, enum asset_e);
 void particleEmitter_setPosition(ParticleEmitter *, f32[3]);
 ParticleEmitter *partEmitMgr_newEmitter(u32);
 void func_802BB3DC(s32, f32, f32);
-void __spawnQueue_add_4(GenFunction_4, s32, s32, s32, s32);
+void __spawnQueue_add_4(GenFunction_4, intptr_t, intptr_t, intptr_t, intptr_t);
 Actor *spawnQueue_actor_f32(enum actor_e actor_id, s32 x, s32 y, s32 z);
 void sfxsource_freeSfxsourceByIndex(u8);
